+++ CHANGELOG.md
@@ -7,16 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `rebar3 grisp package list` can now take a `--hash` flag that prints all hash versions of a package
+
 ## [2.1.0] - 2022-02-14
 
 ### Added
 
-- `rebar3 grisp package list` task that lists pre-built OTP package versions.
+- `rebar3 grisp package list` task that lists pre-built OTP package versions
 
 ### Changed
 
 - `deploy` task no longer requires specifying release name and version if there
-  is only one release.
+  is only one release
 
 ### Fixed